Implemented basic summaries and announcements.
