+++ lib/components/index.ts
@@ -1,2 +1,4 @@
+export * from './aria_live';
 export * from './dialog';
+export * from './paracomponent';
 export * from './settings';
