+++ lib/control_panel/dialogs/index.ts
@@ -1,3 +1,2 @@
-export * from './history_dialog';
 export * from './soni_dialog';
 export * from './advanced_controls_dialog';
