+++ lib/control_panel/setting_control_container.ts
@@ -1,5 +1,5 @@
 
-import { ParaComponent } from '../paracomponent';
+import { ParaComponent } from '../components';
 import { logging } from '../common/logger';
 import '../components';
