+++ lib/parachart/parachart.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.*/
 
 import { logging } from '../common/logger';
-import { ParaComponent } from '../paracomponent';
+import { ParaComponent } from '../components';
 import { ParaController } from '../paracontroller';
 import { AllSeriesData } from '@fizz/paramanifest'
 import { DeepReadonly, Settings, SettingsInput } from "../store/settings_types";
