+++ lib/store/series_properties.ts
@@ -14,7 +14,7 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.*/
 
-import { type DataSymbolType } from '../view_temp/symbol';
+import { type DataSymbolType } from '../view/symbol';
 import { ParaStore } from './parastore';
 
 export class SeriesPropertyManager {
