+++ lib/view/bar.ts
@@ -385,7 +385,7 @@ export class Bar extends XYDatapointView {
   }
 
   content() {
-    const visitedScale = this.paraview.store.visitedDatapoints.includes(this._id)
+    const visitedScale = this.paraview.store.isVisited(this.seriesKey, this.index)
       ? this.chart.settings.highlightScale : 1;
     const styles = {
       strokeWidth: this.chart.settings.lineWidth*visitedScale
