+++ lib/view/base_view.ts
