+++ lib/view/chartlayermanager.ts
