+++ lib/view/document_view.ts
