+++ lib/view/pointchart.ts
