+++ lib/view/rule.ts
