+++ lib/view/scatter.ts
