+++ lib/view/serieslabelstrip.ts
