+++ lib/view/signals.ts
