+++ lib/view/symbol.ts
