+++ lib/view/tickstrip.ts
